Move toObject method to MatrixValue+Object.swift
Refactored the toObject(in:) method from MatrixValue.swift into a new MatrixValue+Object.swift extension file. This improves code organization by separating object conversion logic from the main MatrixValue implementation.
